Add some pause screen data, link pause_screen_map.c

+++ include/types.h
@@ -28,7 +28,7 @@ typedef volatile signed int vi32;
 
 #define INT_MIN (-2147483648)
 #define INT_MAX (2147483647)
-#define UINT_MAX (4294967295)
+#define UINT_MAX ((u32)4294967295)
 
 #define BLOCK_SIZE 0x40
 #define HALF_BLOCK_SIZE (BLOCK_SIZE / 2)
